+++ pyscandir/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.3] - 2024-10-24
+
+### Improved
+
+- Update dependencies.
+- Add support for Python 3.13.
+- Fix continuous integration on github for Linux and Windows.
+- Add support for macos-14 on ARM64.
+
 ## [2.7.2] - 2024-07-09
 
 ### Improved